PEPPER-1402 based on PR comments I increased the initial wait time for a response from dss , and also did a major refactoring for our util classes in test that were creating test kits.
