+++ pepper-apis/dsm-core/src/test/java/org/broadinstitute/dsm/service/adminoperation/UpdateKitToLegacyIdServiceTest.java
@@ -11,7 +11,7 @@ import org.broadinstitute.dsm.db.KitRequestShipping;
 import org.broadinstitute.dsm.db.dao.ddp.instance.DDPInstanceDao;
 import org.broadinstitute.dsm.db.dto.ddp.instance.DDPInstanceDto;
 import org.broadinstitute.dsm.db.dto.ddp.participant.ParticipantDto;
-import org.broadinstitute.dsm.kits.TestKitUtil;
+import org.broadinstitute.dsm.kits.PepperKitUtil;
 import org.broadinstitute.dsm.model.elastic.Dsm;
 import org.broadinstitute.dsm.model.elastic.search.ElasticSearchParticipantDto;
 import org.broadinstitute.dsm.service.elastic.ElasticSearchService;
@@ -33,7 +33,7 @@ public class UpdateKitToLegacyIdServiceTest extends DbAndElasticBaseTest {
     private static String ddpParticipantId = "PT_SAMPLE_QUEUE_TEST";
     private static int participantCounter = 0;
     private static Pair<ParticipantDto, String> legacyParticipantPair;
-    private static TestKitUtil testKitUtil;
+    private static PepperKitUtil pepperKitUtil;
     private static final String newCollaboratorSampleId = "NEW_SAMPLE_ID";
     private static final String newCollaboratorParticipantId = "NEW_PARTICIPANT_ID";
 
@@ -47,8 +47,9 @@ public class UpdateKitToLegacyIdServiceTest extends DbAndElasticBaseTest {
     @BeforeClass
     public static void doFirst() {
         esIndex = ElasticTestUtil.createIndex(instanceName, "elastic/lmsMappings.json", null);
-        testKitUtil = new TestKitUtil(instanceName, instanceName, "UpdateCollab", instanceName, "SALIVA", null, esIndex);
-        testKitUtil.setupInstanceAndSettings();
+        pepperKitUtil = new PepperKitUtil(instanceName, instanceName, "UpdateCollab", instanceName, "SALIVA", null, esIndex,
+                null);
+        pepperKitUtil.setupInstanceAndSettings();
         ddpInstanceDto = ddpInstanceDao.getDDPInstanceByInstanceName(instanceName).orElseThrow();
         legacyParticipantPair = TestParticipantUtil.createLegacyParticipant(ddpParticipantId, participantCounter++, ddpInstanceDto,
                 shortId, legacyShortId);
@@ -59,9 +60,9 @@ public class UpdateKitToLegacyIdServiceTest extends DbAndElasticBaseTest {
                 .withBspCollaboratorParticipantId(oldCollaboratorParticipantId)
                 .withDdpKitRequestId(ddpKitRequestId)
                 .withKitTypeName("SALIVA")
-                .withKitTypeId(String.valueOf(testKitUtil.kitTypeId)).build();
+                .withKitTypeId(String.valueOf(pepperKitUtil.getKitTypeId())).build();
 
-        String dsmKitRequestId = testKitUtil.createKitRequestShipping(kitRequestShipping, DDPInstance.from(ddpInstanceDto), "100");
+        String dsmKitRequestId = pepperKitUtil.createKitRequestShipping(kitRequestShipping, DDPInstance.from(ddpInstanceDto), "100");
         createdKits.add(dsmKitRequestId);
     }
 
@@ -69,8 +70,8 @@ public class UpdateKitToLegacyIdServiceTest extends DbAndElasticBaseTest {
     public static void tearDown() {
         participants.forEach(participantDto ->
                 TestParticipantUtil.deleteParticipant(participantDto.getRequiredParticipantId()));
-        createdKits.forEach(dsmKitRequestId -> testKitUtil.deleteKitRequestShipping((Integer.parseInt(dsmKitRequestId))));
-        testKitUtil.deleteGeneratedData();
+        createdKits.forEach(dsmKitRequestId -> pepperKitUtil.deleteKitRequestShipping((Integer.parseInt(dsmKitRequestId))));
+        pepperKitUtil.deleteGeneratedData();
         ddpInstanceDao.delete(ddpInstanceDto.getDdpInstanceId());
         ElasticTestUtil.deleteIndex(esIndex);
     }
@@ -110,8 +111,8 @@ public class UpdateKitToLegacyIdServiceTest extends DbAndElasticBaseTest {
                 .withBspCollaboratorParticipantId(collaboratorParticipantId)
                 .withDdpKitRequestId("KIT_REQ_ID_DUP_KIT1")
                 .withKitTypeName("SALIVA")
-                .withKitTypeId(String.valueOf(testKitUtil.kitTypeId)).build();
-        String dsmKitRequestId = testKitUtil.createKitRequestShipping(kitRequestShipping, DDPInstance.from(ddpInstanceDto), "100");
+                .withKitTypeId(String.valueOf(pepperKitUtil.getKitTypeId())).build();
+        String dsmKitRequestId = pepperKitUtil.createKitRequestShipping(kitRequestShipping, DDPInstance.from(ddpInstanceDto), "100");
         createdKits.add(dsmKitRequestId);
         String dupCollaboratorSampleId = "DUP_COLLABORATOR_SAMPLE_ID";
         KitRequestShipping kitRequestShipping2 =  KitRequestShipping.builder()
@@ -120,8 +121,8 @@ public class UpdateKitToLegacyIdServiceTest extends DbAndElasticBaseTest {
                 .withBspCollaboratorParticipantId(collaboratorParticipantId)
                 .withDdpKitRequestId("KIT_REQ_ID_DUP_KIT2")
                 .withKitTypeName("SALIVA")
-                .withKitTypeId(String.valueOf(testKitUtil.kitTypeId)).build();
-        String dsmKitRequestId2 = testKitUtil.createKitRequestShipping(kitRequestShipping2, DDPInstance.from(ddpInstanceDto), "100");
+                .withKitTypeId(String.valueOf(pepperKitUtil.getKitTypeId())).build();
+        String dsmKitRequestId2 = pepperKitUtil.createKitRequestShipping(kitRequestShipping2, DDPInstance.from(ddpInstanceDto), "100");
         createdKits.add(dsmKitRequestId2);
 
         UpdateKitToLegacyIdsRequest duplicateSampleIdRequest = new UpdateKitToLegacyIdsRequest(collaboratorSampleId,
